requirements.txt
Flask==2.3.3
openai==1.3.0
numpy==1.24.3
faiss-cpu==1.7.4
python-dotenv==1.0.0 